Move wrapper type instances to new module
