+++ src/Proto3/Suite/DotProto/Generate/Swagger.hs
@@ -29,17 +29,13 @@ import           Data.Aeson                      (Value (String), ToJSONKey,
 import qualified Data.Aeson                      as Aeson
 import           Data.Aeson.Encode.Pretty        (encodePretty)
 import qualified Data.ByteString                 as B
-import qualified Data.ByteString.Lazy            as BL
 import qualified Data.ByteString.Lazy.Char8      as LC8
-import           Data.Functor                    ((<&>))
 import           Data.Hashable                   (Hashable)
 import           Data.HashMap.Strict.InsOrd      (InsOrdHashMap)
 import qualified Data.HashMap.Strict.InsOrd
 import           Data.Map                        (Map)
 import           Data.Swagger
-import           Data.Swagger.Declare            (Declare)
 import qualified Data.Text                       as T
-import qualified Data.Text.Lazy                  as TL
 import           Data.Proxy
 import qualified Data.Vector                     as V
 import           GHC.Exts                        (Proxy#, proxy#)
@@ -47,6 +43,7 @@ import           GHC.Int
 import           GHC.Word
 import           Proto3.Suite                    (Enumerated (..), Finite (..),
                                                   Fixed (..), Named (..), enumerate)
+import           Proto3.Suite.DotProto.Generate.Swagger.Wrappers ()
 
 -- | Convenience re-export so that users of generated code don't have to add
 --   an explicit dependency on @insert-ordered-containers@
@@ -89,66 +86,6 @@ instance {-# OVERLAPPING #-} (ToJSONKey k, ToSchema k) => ToSchema (OverrideToSc
         & type_ .~ SwaggerObject
 #endif
         & additionalProperties ?~ AdditionalPropertiesSchema (Inline byteSchema)
-
-#ifdef SWAGGER_WRAPPER_FORMAT
--- | Wrapped Type Schemas
-
-setFormat
-  :: T.Text
-  -> Declare (Definitions Schema) NamedSchema
-  -> Declare (Definitions Schema) NamedSchema
-setFormat formatValue namedSchema =
-  namedSchema
-#if MIN_VERSION_swagger2(2,4,0)
-    <&> schema . paramSchema . format ?~ formatValue
-#else
-    <&> schema . paramSchema . format .~ formatValue
-#endif
-
-declareWrapperNamedSchema
-  :: forall a
-   . ToSchema a
-  => T.Text
-  -> Proxy (OverrideToSchema a)
-  -> Declare (Definitions Schema) NamedSchema
-declareWrapperNamedSchema formatValue _ =
-  setFormat formatValue (declareNamedSchema (Proxy :: Proxy a))
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Double)) where
-  declareNamedSchema = declareWrapperNamedSchema "DoubleValue"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Float)) where
-  declareNamedSchema = declareWrapperNamedSchema "FloatValue"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Int64)) where
-  declareNamedSchema = declareWrapperNamedSchema "Int64Value"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Word64)) where
-  declareNamedSchema = declareWrapperNamedSchema "UInt64Value"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Int32)) where
-  declareNamedSchema = declareWrapperNamedSchema "Int32Value"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Word32)) where
-  declareNamedSchema = declareWrapperNamedSchema "UInt32Value"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe Bool)) where
-  declareNamedSchema = declareWrapperNamedSchema "BoolValue"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe T.Text)) where
-  declareNamedSchema = declareWrapperNamedSchema "StringValue"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe TL.Text)) where
-  declareNamedSchema = declareWrapperNamedSchema "StringValue"
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe B.ByteString)) where
-  declareNamedSchema _ =
-    setFormat "BytesValue" (pure (NamedSchema Nothing byteSchema))
-
-instance {-# OVERLAPPING #-} ToSchema (OverrideToSchema (Maybe BL.ByteString)) where
-  declareNamedSchema _ =
-    setFormat "BytesValue" (pure (NamedSchema Nothing byteSchema))
-#endif
 
 {-| This is a convenience function that uses type inference to select the
     correct instance of `ToSchema` to use for fields of a message
